ISAICP-4518: Steps that cannot run in batch process should end the PHP script execution.
