+++ web/modules/custom/joinup_federation/src/Plugin/pipeline/Step/AdmsValidation.php
@@ -6,6 +6,8 @@ namespace Drupal\joinup_federation\Plugin\pipeline\Step;
 
 use Drupal\adms_validator\AdmsValidatorInterface;
 use Drupal\joinup_federation\JoinupFederationStepPluginBase;
+use Drupal\pipeline\Plugin\PipelineStepWithRedirectResponseTrait;
+use Drupal\pipeline\Plugin\PipelineStepWithResponseInterface;
 use Drupal\rdf_entity\Database\Driver\sparql\Connection;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -17,7 +19,9 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  *   label = @Translation("ADMS Validation"),
  * )
  */
-class AdmsValidation extends JoinupFederationStepPluginBase {
+class AdmsValidation extends JoinupFederationStepPluginBase implements PipelineStepWithResponseInterface {
+
+  use PipelineStepWithRedirectResponseTrait;
 
   /**
    * The ADMS validator service.
